requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
google-api-core==2.24.2
google-auth==2.38.0
google-cloud==0.34.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.30.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
plotly==6.0.0
plotly-express==0.4.1
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
retrying==1.3.4
rsa==4.9
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
statsmodels==0.14.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0
